Refactor: use commons package with main abstraction
