+++ src/main/java/com/joutvhu/dynamic/freemarker/FreemarkerDynamicQueryTemplate.java
@@ -1,6 +1,6 @@
-package com.joutvhu.dynamic.commons.freemarker;
+package com.joutvhu.dynamic.freemarker;
 
-import com.joutvhu.dynamic.jpa.DynamicQueryTemplate;
+import com.joutvhu.dynamic.commons.DynamicQueryTemplate;
 import freemarker.template.Template;
 import lombok.RequiredArgsConstructor;
