--- src/main/java/com/joutvhu/dynamic/jpa/DynamicQueryTemplate.java
@@ -1,5 +1,0 @@
-package com.joutvhu.dynamic.jpa;
-
-public interface DynamicQueryTemplate<T> {
-    T getTemplate();
-}
