--- src/main/java/com/joutvhu/dynamic/jpa/DynamicQueryTemplateHandler.java
@@ -1,4 +1,0 @@
-package com.joutvhu.dynamic.jpa;
-
-public interface DynamicQueryTemplateHandler<T> extends DynamicQueryTemplateProvider<T>, DynamicQueryTemplateProcessor<T> {
-}
