--- src/main/java/com/joutvhu/dynamic/jpa/DynamicQueryTemplateProcessor.java
@@ -1,7 +1,0 @@
-package com.joutvhu.dynamic.jpa;
-
-import java.util.Map;
-
-public interface DynamicQueryTemplateProcessor<T> {
-    String processTemplate(DynamicQueryTemplate<T> template, Map<String, Object> params);
-}
