--- src/main/java/com/joutvhu/dynamic/jpa/DynamicQueryTemplateProvider.java
@@ -1,6 +1,0 @@
-package com.joutvhu.dynamic.jpa;
-
-public interface DynamicQueryTemplateProvider<T> {
-    DynamicQueryTemplate<T> createTemplate(String name, String query);
-    DynamicQueryTemplate<T> findTemplate(String name);
-}
