+++ src/main/java/com/joutvhu/dynamic/jpa/query/DynamicJpaQueryMethod.java
@@ -1,8 +1,8 @@
 package com.joutvhu.dynamic.jpa.query;
 
 import com.joutvhu.dynamic.jpa.DynamicQuery;
-import com.joutvhu.dynamic.jpa.DynamicQueryTemplate;
-import com.joutvhu.dynamic.jpa.DynamicQueryTemplateProvider;
+import com.joutvhu.dynamic.commons.DynamicQueryTemplate;
+import com.joutvhu.dynamic.commons.DynamicQueryTemplateProvider;
 import com.joutvhu.dynamic.commons.util.ApplicationContextHolder;
 import org.springframework.core.annotation.AnnotatedElementUtils;
 import org.springframework.core.annotation.AnnotationUtils;
