+++ src/test/java/com/joutvhu/dynamic/freemarker/DirectiveTest.java
@@ -1,7 +1,6 @@
 
 package com.joutvhu.dynamic.freemarker;
 
-import com.joutvhu.dynamic.commons.freemarker.FreemarkerTemplateConfiguration;
 import freemarker.template.Configuration;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
