+++ src/test/java/com/joutvhu/dynamic/jpa/JpaDynamicApplication.java
@@ -1,7 +1,9 @@
 package com.joutvhu.dynamic.jpa;
 
-import com.joutvhu.dynamic.commons.freemarker.FreemarkerDynamicQueryTemplateHandler;
+import com.joutvhu.dynamic.commons.DynamicQueryTemplateHandler;
+import com.joutvhu.dynamic.freemarker.FreemarkerDynamicQueryTemplateHandler;
 import com.joutvhu.dynamic.jpa.support.DynamicJpaRepositoryFactoryBean;
+import freemarker.template.Template;
 import org.hibernate.cfg.AvailableSettings;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.SpringApplication;
@@ -82,7 +84,7 @@ public class JpaDynamicApplication {
     }
 
     @Bean
-    public DynamicQueryTemplateHandler dynamicQueryTemplates() {
+    public DynamicQueryTemplateHandler<Template> dynamicQueryTemplates() {
         FreemarkerDynamicQueryTemplateHandler templateHandler = new FreemarkerDynamicQueryTemplateHandler();
         templateHandler.setSuffix(".dsql");
         return templateHandler;
